Welcome page UI

+++ src/main/java/Entities/Farmer.java
@@ -7,18 +7,19 @@ public class Farmer extends User {
     private final ArrayList<RequestStructure> current_requests;
     private final ArrayList<Offer> offer_history;
 
-    private int prefPrice; // Int from 0-10 detailing preference for exact or better price (10 is rigid price)
-    private int prefExposure; // Int from 0-10 detailing preference for high exposure (10 is maximum exposure)
-    private int prefSpeed; // Int from 0-10 detailing preference for fast pickup (10 is fastest pickup)
-    private int prefCarbon; // Int from 0-10 detailing preference for low-carbon distribution (10 is lowest emissions)
+    private double prefPrice; // Int from 0-10 detailing preference for exact or better price (10 is rigid price)
+    private double prefExposure; // Int from 0-10 detailing preference for high exposure (10 is maximum exposure)
+    private double prefSpeed; // Int from 0-10 detailing preference for fast pickup (10 is fastest pickup)
+    private double prefCarbon; // Int from 0-10 detailing preference for low-carbon distribution (10 is lowest emissions)
 
-    public Farmer(String farmer_name, String farmer_address){
+    public Farmer(String farmer_name, String farmer_address) {
         super(farmer_name, farmer_address);
         this.current_requests = new ArrayList<>();
         this.offer_history = new ArrayList<>();
+
     }
 
-    public void add_request(Request request){
+    public void add_request(Request request) {
         this.current_requests.add(request);
     }
 
@@ -26,11 +27,11 @@ public class Farmer extends User {
         this.current_requests.remove(request);
     }
 
-    public ArrayList<RequestStructure> getCurrent_requests(){
+    public ArrayList<RequestStructure> getCurrent_requests() {
         return this.current_requests;
     }
 
-    public void add_offer(Offer offer){
+    public void add_offer(Offer offer) {
         this.offer_history.add(offer);
     }
 
@@ -38,42 +39,42 @@ public class Farmer extends User {
         this.offer_history.remove(offer);
     }
 
-    public ArrayList<Offer> getOffer_history(){
+    public ArrayList<Offer> getOffer_history() {
         return this.offer_history;
     }
 
     //=========================================================================
 
-    public void setPrefPrice(int prefPrice) {
+    public void setPrefPrice(double prefPrice) {
         this.prefPrice = prefPrice;
     }
 
-    public void setPrefExposure(int prefExposure) {
+    public void setPrefExposure(double prefExposure) {
         this.prefExposure = prefExposure;
     }
 
-    public void setPrefSpeed(int prefSpeed) {
+    public void setPrefSpeed(double prefSpeed) {
         this.prefSpeed = prefSpeed;
     }
 
-    public void setPrefCarbon(int prefCarbon) {
+    public void setPrefCarbon(double prefCarbon) {
         this.prefCarbon = prefCarbon;
     }
 
-    public int getPrefPrice(){
+    public double getPrefPrice() {
         return prefPrice;
     }
 
-    public int getPrefExposure(){
+    public double getPrefExposure() {
         return prefExposure;
     }
 
-    public int getPrefSpeed(){
+    public double getPrefSpeed() {
         return prefSpeed;
     }
 
-    public int getPrefCarbon(){
+    public double getPrefCarbon() {
         return prefCarbon;
     }
-  
+
 }
